+++ src/tcode_api/schemas/commands/union.py
@@ -19,6 +19,21 @@ from .calibrate_labware_well_depth.latest import (
 from .calibrate_tool.latest import (
     CALIBRATE_TOOL,
 )
+from .camera_configure.latest import (
+    CAMERA_CONFIGURE,
+)
+from .camera_set_enabled.latest import (
+    CAMERA_SET_ENABLED,
+)
+from .camera_start_recording.latest import (
+    CAMERA_START_RECORDING,
+)
+from .camera_stop_recording.latest import (
+    CAMERA_STOP_RECORDING,
+)
+from .camera_take_picture.latest import (
+    CAMERA_TAKE_PICTURE,
+)
 from .comment.latest import COMMENT
 from .create_labware.latest import CREATE_LABWARE
 from .delete_labware.latest import DELETE_LABWARE
@@ -58,6 +73,11 @@ TCode = Annotated[
     | CALIBRATE_LABWARE_HEIGHT
     | CALIBRATE_LABWARE_HOLDER
     | CALIBRATE_TOOL
+    | CAMERA_CONFIGURE
+    | CAMERA_SET_ENABLED
+    | CAMERA_START_RECORDING
+    | CAMERA_STOP_RECORDING
+    | CAMERA_TAKE_PICTURE
     | COMMENT
     | CREATE_LABWARE
     | DELETE_LABWARE